1.5 "Чёрная магия" Spring Boot
