+++ README.md
@@ -4,5 +4,7 @@
 
 Список занятие-работа:
 
-**1.1.** ДЗ по уроку Введение в Spring Framework - spring-homework-1
-**1.3.** ДЗ по уроку Конфигурирование Spring-приложений - spring-homework-2
+**1.1.** ДЗ по уроку Введение в Spring Framework - homework-1
+**1.3.** ДЗ по уроку Конфигурирование Spring-приложений - homework-2
+**1.4** Unit-тестирование - homework-3
+**1.5** "Чёрная магия" Spring Boot - homework-4
