--- spring-homework-2/src/test/resources/exam.properties
@@ -1,2 +1,0 @@
-file.name=/questions.csv
-pass.percent=75
