+++ spring-homework-4/src/main/java/ru/otus/exam/domain/StudentCard.java
