+++ spring-homework-4/src/main/java/ru/otus/exam/service/PrinterService.java
@@ -1,0 +1,10 @@
+package ru.otus.exam.service;
+
+import org.springframework.lang.Nullable;
+
+public interface PrinterService {
+
+    void print(String code);
+
+    void print(String code, @Nullable Object... args);
+}
