+++ spring-homework-4/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+pass.percent:75
+file.name:questions
+exam.locale:en-EN
