+++ spring-homework-4/src/main/resources/banner.txt
@@ -1,0 +1,7 @@
+${Ansi.BLUE}
+    ,------.                                   ,---.     ,--.                      ,--.              ,--.
+    |  .---' ,--.  ,--.  ,--,--. ,--,--,--.   '   .-'  ,-'  '-.  ,--,--. ,--.--. ,-'  '-.  ,---.   ,-|  |${Ansi.MAGENTA}
+    |  `--,   \  `'  /  ' ,-.  | |        |   `.  `-.  '-.  .-' ' ,-.  | |  .--' '-.  .-' | .-. : ' .-. |
+    |  `---.  /  /.  \  \ '-'  | |  |  |  |   .-'    |   |  |   \ '-'  | |  |      |  |   \   --. \ `-' |${Ansi.BRIGHT_MAGENTA}
+    `------' '--'  '--'  `--`--' `--`--`--'   `-----'    `--'    `--`--' `--'      `--'    `----'  `---'
+${Ansi.DEFAULT}
