+++ spring-homework-4/src/main/resources/messages.properties
@@ -1,0 +1,10 @@
+exam.firstname.string=Добро пожаловать! Пожалуйста, укажите свое имя:
+exam.lastname.string=И вашу фамилию:
+exam.congrats.string=Поздравляем, вы успешно завершили тестирование!
+exam.fail.string=Вам нужно больше знаний для прохождения тестирования.
+exam.score.string=Ваши баллы %d%n
+
+control.answer.string=\t%d - %s%n
+control.chose.answer.string:Пожалуйста, выберите ответ:
+control.wrong.answer.string=Используйте для отчета только числа. Пожалуйста, ответьте снова:
+control.students.answer.string=Ваш ответ %d%n
