+++ spring-homework-4/src/main/resources/messages_en_EN.properties
@@ -1,0 +1,10 @@
+exam.firstname.string=Hello! Please, tape your first name:
+exam.lastname.string=And your last name:
+exam.congrats.string=Congrats, you pass!
+exam.fail.string=You need to study more.
+exam.score.string=Your score is %d%n
+
+control.answer.string=\t%d - %s%n
+control.chose.answer.string=Please, chose your answer:
+control.wrong.answer.string=You should use only numbers. Please answer again:
+control.students.answer.string=Your answer is %d%n
