+++ spring-homework-4/src/main/resources/questions.csv
@@ -1,0 +1,1 @@
+Могу я припарковаться здесь?;Извините, это был я.;Это то же самое место.;Только на пол часа.;3Какой цвет вы выбрали для детской комнаты?;Надеюсь, это было верно.;Мы не решили.;Это было не слишком сложно.;2Я не могу понять это письмо.;Может, тебе нужна помощь?;Разве ты не знаешь?;Я полагаю, ты можешь.;1Два билета на завтрашнюю ночь, пожалуйста.;Как много ты заплатил?;После полудня и вечером;Я уточню для вас.;3Может нам пойти в спортзал?;Я слишком устал.;Это очень хорошо.;Вовсе нет.;1
