+++ spring-homework-4/src/test/java/ru/otus/exam/ExamApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.otus.exam;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ExamApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
