+++ spring-homework-4/src/test/java/ru/otus/exam/service/ExamServiceTest.java
@@ -1,0 +1,68 @@
+package ru.otus.exam.service;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.context.MessageSource;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.PrintStream;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+@DisplayName("Сервис запуска тестирования")
+class ExamServiceTest {
+
+    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
+    private final PrintStream originalOut = System.out;
+
+    private ExamService examineService;
+
+    private PrinterService printerServiceMock;
+
+    private ControlService controlServiceMock;
+    private ReaderService readerService;
+
+    @BeforeEach
+    private void setUp() {
+        System.setOut(new PrintStream(outContent));
+
+        readerService = mock(ReaderService.class);
+
+        controlServiceMock = mock(ControlService.class);
+
+        printerServiceMock = mock(PrinterService.class);
+
+        examineService = new ExamServiceImpl(controlServiceMock, readerService, printerServiceMock);
+    }
+
+    @AfterEach
+    public void restoreStreams() {
+        System.setOut(originalOut);
+    }
+
+    @Test
+    @DisplayName("Сообщение верное при прохождении порога верных ответов")
+    void examPass() throws IOException {
+        when(readerService.read()).thenReturn("Ivan","Ivanov");
+        when(controlServiceMock.control(any())).thenReturn(true);
+
+        examineService.exam();
+
+        assertTrue(outContent.toString().contains("Congrats, you pass!"));
+    }
+
+    @Test
+    @DisplayName("Сообщение верное при недостижении порога верных ответов")
+    void examNotPass() throws IOException {
+        when(readerService.read()).thenReturn("Ivan","Ivanov");
+        when(controlServiceMock.control(any())).thenReturn(false);
+
+        examineService.exam();
+
+        assertTrue(outContent.toString().contains("You need to study more."));
+    }
+}
